PDFBOX-5660: Length entry can be long instead of int
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1932950 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/cos/COSStream.java
@@ -374,7 +374,7 @@ public class COSStream extends COSDictionary implements Closeable
                                             "COSStream. It must be closed before querying the " +
                                             "length of this COSStream.");
         }
-        return getInt(COSName.LENGTH, 0);
+        return getLong(COSName.LENGTH, 0);
     }
 
     /**
